[Refactor] 훅 및 쿼리 구조 개선 - 불필요한 훅 제거 및 쿼리 통합
