--- src/hooks/useSelectedStock.ts
@@ -1,9 +1,0 @@
-import { getSelectedStock } from '@/api/stock.api';
-import { useSuspenseQuery } from '@tanstack/react-query';
-
-export const useSelectedStock = () => {
-    return useSuspenseQuery({
-        queryKey: ['selectedStock'],
-        queryFn: () => getSelectedStock(),
-    });
-};
