--- src/hooks/useStocks.ts
@@ -1,11 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { getStocks } from '@/api/stock.api';
-
-export const useStocks = () => {
-    const { data, isLoading, error } = useQuery({
-        queryKey: ['stocks'],
-        queryFn: () => getStocks(),
-    });
-
-    return { data, isLoading, error };
-};
